smartrentals_mvp/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.32
pydantic==2.8.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
